Fixed issue with a blank arg to ovftool and bad string formatting
Fixes an issue where if a compression level was not specified, the args passed
to ovftool would include a blank string arg, which it would take as the vmx
source, then the vmx would be the destination, and the destination arg triggered
an error.

Also fixed a few places that had string formating without including the variable
in the format string.